requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Django==4.0.5
django-phonenumber-field==6.2.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
kombu==5.2.4
phonenumbers==8.12.50
prompt-toolkit==3.0.29
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
